Stock pool: point-in-time CSI 300 constituents, delisted stocks, financials by notice date
- tools/build_hs300.py rebuilds CSI 300 membership since 2005-07 from CSIndex
  announcements (HTML tables, xls/xlsx, pdf) plus hand-checked ad-hoc changes
  (tools/hs300_adhoc.json); every step must hold exactly 300 names and the
  result must equal the official current list. Output: docs/data/hs300_members.json
  (67 adjustments, 941 stocks, merger delisting dates)
- docs/stockpool.js: fetch prices (hfq + raw with turnover) and F10 financials
  (includes delisted companies) per stock for its membership window, store in
  IndexedDB with resume, assemble data + universe/fundamentals/delisted meta
- New "个股池" data source tab; delisting recovery setting; PIT benchmark label;
  signal list and series stats trimmed for large pools
- Memory for ~560-stock pools: shared backtester across tabs, lazy forward
  fills, compact typed arrays (842MB -> 491MB JS heap in the e2e run)
- Tests for parsers, the committed membership file (300 names every month),
  pool assembly and new eastmoney parsers; weekly job reports stale membership

+++ docs/research.js
@@ -717,7 +717,7 @@
         .filter((r) => r.notice && r.report)
         .slice()
         .sort((x, y) => (x.notice < y.notice ? -1 : x.notice > y.notice ? 1 : x.report < y.report ? -1 : 1));
-      const col = () => new Float64Array(n).fill(NaN);
+      const col = () => new Float32Array(n).fill(NaN); // 财务指标单精度足够，省内存
       const out = { epsTTM: col(), bps: col(), roe: col(), revYoy: col(), profitYoy: col() };
       const known = new Map();
       let latest = null, j = 0, ttm = NaN;
